requirements.txt
mcp
psycopg2-binary
anthropic
